requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.1
defusedxml==0.7.1
Django==3.2.19
django-allauth==0.54.0
django-crispy-forms==1.14.0
django-filter==2.4.0
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2018.5
requests==2.31.0
requests-oauthlib==1.3.1
shortuuid==1.0.11
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3